+++ pengucoach/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.1.0-alpha.32"
+__version__ = "0.1.0-alpha.33"